Add types to `withEditMode`
