+++ assets/js/blocks/featured-items/types.ts
@@ -1,6 +1,18 @@
+/**
+ * External dependencies
+ */
+import { Block, BlockEditProps } from '@wordpress/blocks';
+
+export type EditorBlock< T > = Block< T > & BlockEditProps< T >;
+
 export interface Coordinates {
 	x: number;
 	y: number;
+}
+
+export interface GenericBlockUIConfig {
+	icon: JSX.Element;
+	label: string;
 }
 
 export type ImageFit = 'cover' | 'none';
